### 2025-09-22 #9042 @editorialbot
[PRE REVIEW]: traversability_generator3d: A C++ library for 3D traversability estimation from MLS maps
**Submitting author:** @haider8645 (<a href="http://orcid.org/0009-0008-1199-3489">Muhammad Haider Khan Lodhi</a>)
**Repository:** https://github.com/dfki-ric/traversability_generator3d
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
