### 2025-09-23 14f45a8
Remove unused patch type
