### 2025-10-02 #9042 @haider8645
@logological , would this paper be of interest to you?
