### 2025-11-02 #9042 @logological
> [@logological](https://github.com/logological) , would this paper be of interest to you?

@haider8645 Apologies for not having responded to this question, but it seems I didn't get an e-mail notification for this mention!  I'm happy to see you've since found an editor!
