### 2025-11-03 ac4c3e0
Remove unused code
