### 2025-11-05 #9042 @boisgera
**Update:** I have one potential reviewer to begin with if he can evaluate the software for Ubuntu 24.04 / amd64.

But so far I have not been able to build the software myself for this platform in a reasonable amount of time (I am trying to build it in an isolated [pixi](https://prefix.dev/) environment but it *should* not matter; I am not familiar with the Ruby tooling which is used but I tried my best...).

I'll document my failed process when I get a bit more time in a project issue. I think we'll need to get this issue solved to get the commitment of this reviewer (which is fair IMHO).
