### 2025-11-05 c25f0c7
Move library viz dep to executable DEPS_QT
