### 2025-11-06 #9042 @haider8645
> No, the `main` branch. Is this significant?

The paper branch has installation scripts and related instructions to automatically build and install dependencies and the library
