### 2025-11-06 dfki-ric/traversability_generator3d#11 @boisgera
Install
JOSS context: https://github.com/openjournals/joss-reviews/issues/9042#issuecomment-3476394379

The [installation instructions](https://github.com/dfki-ric/traversability_generator3d/tree/paper?tab=readme-ov-file#automatic-install-of-dependencies--build) (with Ubuntu 24.04 my machine and fresh install) fail with

```
Cloning into 'traversability_generator3d'...
remote: Enumerating objects: 628, done.
remote: Total 628 (delta 20), reused 31 (delta 13), pack-reused 569 (from 1)
Resolving deltas: 100% (382/382), done.
-- The C compiler identification is GNU 13.3.0
-- The CXX compiler identification is GNU 13.3.0
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /usr/bin/cc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /usr/bin/c++ - skipped
-- Detecting CXX compile features
CMake Warning at CMakeLists.txt:7 (find_package):
  By not providing "FindRock.cmake" in CMAKE_MODULE_PATH this project has
  asked CMake to find a package configuration file provided by "Rock", but
  CMake did not find one.

  Could not find a package configuration file provided by "Rock" with any of
  the following names:

    RockConfig.cmake
    rock-config.cmake

  Add the installation prefix of "Rock" to CMAKE_PREFIX_PATH or set
  "Rock_DIR" to a directory containing one of the above files.  If "Rock"
  provides a separate development package or SDK, be sure it has been
  installed.

-- ENABLE_DEBUG_DRAWINGS is set to OFF. Skipped!
CMake Error at CMakeLists.txt:19 (rock_init):
  Unknown CMake command "rock_init".

-- Configuring incomplete, errors occurred!
make: *** No rule to make target 'install'.  Stop.
bash: install/env.sh: No such file or directory
```

The README mentions that this process auto-installs the dependencies ; Therefore I have merely installed cmake, git and g++ beforehand.
