### 2025-11-13 #9042 @boisgera
Hi @haider8645,

I have found another researcher willing to review your submission provided that I can build and run the tests beforehand on a Ubuntu 22.04 platform.

Did you by any chance identify what was the issue in https://github.com/dfki-ric/traversability_generator3d/issues/11?

Best regards,

Sébastien
