### 2025-11-13 dfki-ric/traversability_generator3d#11 @haider8645
@boisgera thanks for trying it out. Actually the branch flag was missing in the clone command. Can you try the following steps and check if it builds for you.

```
git clone https://github.com/dfki-ric/traversability_generator3d.git -b paper
cd traversability_generator3d
mkdir build && cd build
cmake -DINSTALL_DEPS=ON -DRUN_TESTS=ON -DCMAKE_INSTALL_PREFIX=./install ..
make install
source install/env.sh
```
