### 2025-11-13 #9042 @haider8645
> Did you by any chance identify what was the issue in [dfki-ric/traversability_generator3d#11](https://github.com/dfki-ric/traversability_generator3d/issues/11)?

Thanks for the reminder. Actually I did not get a notification for the newly created issue hence the delayed response!
