### 2025-11-13 dfki-ric/traversability_generator3d#11 @boisgera
There is some progress! But it ends with the following error:

```
-- Checking for module 'base-types'
--   Package 'base-types', required by 'virtual:world', not found
CMake Error at /usr/share/cmake-3.28/Modules/FindPkgConfig.cmake:619 (message):
  The following required packages were not found:

   - base-types

Call Stack (most recent call first):
  /usr/share/cmake-3.28/Modules/FindPkgConfig.cmake:841 (_pkg_check_modules_internal)
  build/install/share/rock/cmake/Rock.cmake:334 (pkg_check_modules)
  build/install/share/rock/cmake/Rock.cmake:532 (rock_find_pkgconfig)
  build/install/share/rock/cmake/Rock.cmake:955 (rock_target_definition)
  build/install/share/rock/cmake/Rock.cmake:1053 (rock_library_common)
  src/CMakeLists.txt:9 (rock_library)

-- Configuring incomplete, errors occurred!
make: Nothing to be done for 'install'.
bash: install/env.sh: No such file or directory
```
