### 2025-11-13 dbea4f8
Install OS dependencies also when building doxygen
