### 2025-11-13 d00eed2
Remove unnecessary doc-folder movement
