### 2025-11-13 d5cd732
Remove unneeded os-dependencies for building doc
