### 2025-11-13 dfki-ric/traversability_generator3d#11 @chhtz
Hi @boisgera, we apparently missed an OS dependency (openscenegraph-dev) which made base-types fail to build.
It should be fixed now.

We also set up github actions now, so it should check automatically: https://github.com/dfki-ric/traversability_generator3d/actions/runs/19336662236/job/55313399550

PS: Additionally, we now upload the generated doxygen: https://dfki-ric.github.io/traversability_generator3d/
