### 2025-11-13 f64a39f
Mention OpenSceneGraph dependency in README
