### 2025-11-14 dfki-ric/traversability_generator3d#11 @boisgera
OK, that works for me now, thanks a a lot!
