### 2025-11-14 #9042 @boisgera
Since

  - https://github.com/dfki-ric/traversability_generator3d/issues/11

is solved, I'll register the reviewers and start the review.
