### 2025-11-28 #9410 @haider8645
@boisgera Hi, just wanted to ask about the expected date of start of the review?
