### 2025-12-02 dfki-ric/traversability_generator3d#13 @c-joly
First pass of JOSS review
Joss context: https://github.com/openjournals/joss-reviews/issues/9410
Cc: @boisgera @bstanciulescu

I tried to build and run the code that is on the branch `paper` according to the instructions. It ran smoothly the different steps. After this first stage of review, I have several questions:

 - [x] The GUI has a lot of properties on the right side. Although some of them are obvious, like the "enabled" property of the maps, most of them are difficult to understand. For example, I don't understand the concept of "old data" in the interface. Nothing happens when I try to play with this parameter. Do you have a documentation about all these properties? Moreover, there is a big button **expand** at the bottom of the window, that does not do anything when I click on it. Can you clarify its purpose in the documentation?

  - [x] You claim that your library is also providing a "**soil-aware traversability**". However, the two proposed examples do not embed any information about the soil. Can you add an example that illustrates this functionality?

  - [ ] You claim that there is a **ROS/ROS2 compatibility** with a support of RViz visualization. However, I did not find any information about it in the doc. Can you give more details and some easy examples of this integration?

  - [x] When I run the apps, (GUI and sample example in command line), I would like to **personalize the traversability map to an other robot** (e.g. maximum slope, maximum height step...). Is it possible to have a list of the parameters that are used by the software, and how to change them (with an argument on the command line, config file?)

  - [x] The last command `./build/test/test_travgen utah.ply` just returns me some very high level information in the standard output. However, I don't see any way to get back the traversability map. Is it just a sample code that illustrates the normal usage of the library? In that case, I would expect that the main README provides some **explanations about the associated source files** in order to make a third party user understand fast how works this example. And at least, I would like to see how to read this map in the code.

  - [x] Linked with the previous comment: I haven't found a **documentation about the API** of the library. It does not look like to be built during the installation and I cannot find it on the Github project. Did I miss something?

  - [x] I have some remarks about the behavior of the algorithm. According to my understanding, it starts with an initial position and then try to expand. I tried to apply it on a custom PLY that represents two "cliffs" facing each other (see first screenshot of the GUI).
       - When I click on the left plateau in the GUI, I have one part only that becomes green (2nd screenshot), which is what I expected. When I click on the second plateau (3rd screenshot), both plateau are green. Is there a way to "reset" the traversabilty map in the GUI so I can have only one plateau at a time?
       - When I use the `test_travgen` app with this *ply* file, it ends with:

           ```console
                Loaded point cloud with 20000 points.
                Generated TraversabilityMap3d with 1 nodes.
                Nodes: TRAVERSABLE=0, OBSTACLE=0, FRONTIER=0
           ```
           It looks like the app did not use a good initial point. Thus, it does not find anything in a "silent" way although there are obviously some traversable parts. Can you give some information about the current limitations of this example app `test_travgen`?
