### 2025-12-02 c2ddfba
Add option to reset the TravMap
Signed-off-by: haider8645 <haider_lodhi@hotmail.com>
