### 2025-12-04 1a83490
Add a config to load parameters in GUI
