+++ gui/TraversabilityGenerator3dGui.cpp
@@ -13,6 +13,8 @@
 #include <pcl/common/transforms.h>
 #include <base-logging/Logging.hpp>
 
+#include <yaml-cpp/yaml.h>
+
 #ifdef ENABLE_DEBUG_DRAWINGS
 #include <vizkit3d_debug_drawings/DebugDrawing.hpp>
 #include <vizkit3d_debug_drawings/DebugDrawingColors.hpp>
@@ -69,37 +71,50 @@ void TraversabilityGenerator3dGui::setupUI()
     connect(resetButton, &QPushButton::clicked, this, &TraversabilityGenerator3dGui::resetTravMap);
 }
 
+void TraversabilityGenerator3dGui::loadTravConfigFromYaml(const std::string& file)
+{
+    YAML::Node cfg = YAML::LoadFile(file)["travConfig"];
+
+    LOG_INFO_S << "Loaded Parameters: ";
+    LOG_INFO_S << cfg;
+
+    // doubles
+    travConfig.maxStepHeight            = cfg["maxStepHeight"].as<double>();
+    travConfig.maxSlope                 = cfg["maxSlope"].as<double>();
+    travConfig.inclineLimittingMinSlope = cfg["inclineLimittingMinSlope"].as<double>();
+    travConfig.inclineLimittingLimit    = cfg["inclineLimittingLimit"].as<double>();
+    travConfig.costFunctionDist         = cfg["costFunctionDist"].as<double>();
+    travConfig.minTraversablePercentage = cfg["minTraversablePercentage"].as<double>();
+    travConfig.robotHeight              = cfg["robotHeight"].as<double>();
+    travConfig.robotSizeX               = cfg["robotSizeX"].as<double>();
+    travConfig.robotSizeY               = cfg["robotSizeY"].as<double>();
+    travConfig.distToGround             = cfg["distToGround"].as<double>();
+    travConfig.slopeMetricScale         = cfg["slopeMetricScale"].as<double>();
+    travConfig.gridResolution           = cfg["gridResolution"].as<double>();
+    travConfig.initialPatchVariance     = cfg["initialPatchVariance"] ? cfg["initialPatchVariance"].as<double>() : travConfig.initialPatchVariance;
+
+    // booleans
+    travConfig.allowForwardDownhill     = cfg["allowForwardDownhill"].as<bool>();
+    travConfig.enableInclineLimitting   = cfg["enableInclineLimitting"].as<bool>();
+
+    travConfig.useSoilInformation       = cfg["useSoilInformation"].as<bool>();
+    travConfig.traverseSand             = cfg["traverseSand"].as<bool>();
+    travConfig.traverseRocks            = cfg["traverseRocks"].as<bool>();
+    travConfig.traverseGravel           = cfg["traverseGravel"].as<bool>();
+    travConfig.traverseConcrete         = cfg["traverseConcrete"].as<bool>();
+
+    // enum
+    std::string s = cfg["slopeMetric"].as<std::string>();
+    if     (s == "AVG_SLOPE")      travConfig.slopeMetric = traversability_generator3d::AVG_SLOPE;
+    else if(s == "MAX_SLOPE")      travConfig.slopeMetric = traversability_generator3d::MAX_SLOPE;
+    else if(s == "TRIANGLE_SLOPE") travConfig.slopeMetric = traversability_generator3d::TRIANGLE_SLOPE;
+    else                            travConfig.slopeMetric = traversability_generator3d::NONE;
+}
+
 void TraversabilityGenerator3dGui::setupTravGen(int argc, char** argv)
 {
-    double res = 0.3;
-    if(argc > 2){
-        std::setlocale(LC_ALL, "C");
-        res = atof(argv[2]);
-    }
-
[…]
