### 2025-12-04 dfki-ric/traversability_generator3d#13 @haider8645
Hi, thank you very much for taking the time to build, run, and deeply evaluate the software. Your feedback is extremely helpful. Please find responses to each point below. We will incorporate all necessary clarifications, examples, and documentation updates you mentioned.

I have pushed some changes to please just pull the changes and run ```make install``` in the build folder.

1. GUI properties & unclear parameters ("old data", “expand” button, etc.)

The GUI is based on [gui-vizkit3d](https://github.com/rock-core/gui-vizkit3d?tab=readme-ov-file) which can be used to create 3D visualizations of data. The property "KeepOldData" is a default parameter inserted by the vizkit3d widget class to all sub visualizations. In this case, the parameter is not relevant for our GUI and can be ignored. Additionally, I have removed the Expand button. Previously the traversability map was expanded using the button but it became irrelevant ever since I made it possible to expand based on a mouse click on the MLS surface.

2. Example for soil-aware traversability

Thank you for pointing this out. The feature exists in the library, but the paper branch did not include an example dataset with per-point soil information

I will add:
A minimal script + GUI walkthrough demonstrating how soil layers influence the final traversability map.
Documentation explaining how to provide custom semantic/soil layers.

3. ROS / ROS 2 and RViz integration examples

You are correct that this is not well documented. The visualization of traversability map in ROS2 is provided in the planner package of our stack in [ugv_nav4d_ros2](https://github.com/dfki-ric/ugv_nav4d_ros2). It was done in the planner package because at the time it was the only use-case for the traversability_generator3d. Given that one could have various users in ROS2 for traversability_generator3d alone, I will discuss it and move the visualizations to traversability_generator3d.

4. How to adapt the traversability parameters to another robot

I pushed a change which now includes a yaml parameter file. The parameters are loaded from the yaml in the GUI. You will find it it the folder ```gui/config```. Please install ```sudo apt install libyaml-cpp-dev``` in your workspace. I have added this as a dependency to the source_dependencies as well.

5. test_travgen only prints high-level information

The executable is as you correctly mentioned a walkthrough file with detailed comments in the code on how a user can use traversability_generator3d. Tests were implemented in the [planner library](https://github.com/dfki-ric/ugv_nav4d/blob/main/src/test/test_EnvironmentXYZTheta.cpp) and it is best if I move the tests now to traversability_generator3d.

6. Missing API documentation

The API documentation was linked in the description of the repo but not mentioned in the Readme: https://dfki-ric.github.io/traversability_generator3d/

7. Behavior on the “two plateau cliffs” example

Thank you for providing this useful test case.

(a) GUI: map expansion and resetting

The current GUI accumulates traversability searches unless the map is manually cleared. A reset button can now be used to clear the internal traversability map. Do check it out and let me know if it solves the expansion.

(b) test_travgen finds “1 node only”

The test_travgen assumes that the start position of the robot is at the origin. Based on the visualization, It can be seen that there the origin is in free space, hence no expansion takes place. The 1 node only means that it tried but failed to expand beyond the patch it created at the start location. Again, this script is not really a test but more a detailed walk-through. I will add actual unit tests in the test folder
