### 2025-12-30 73e4449
Add soil samples gui elements
