+++ gui/TraversabilityGenerator3dGui.cpp
@@ -3,6 +3,7 @@
 #include <QPushButton>
 #include <QLabel>
 #include <QComboBox>
+#include <QDoubleSpinBox>
 #include <QVBoxLayout>
 #include <QHBoxLayout>
 #include <vizkit3d/Vizkit3DWidget.hpp>
@@ -12,7 +13,8 @@
 #include <pcl/common/common.h>
 #include <pcl/common/transforms.h>
 #include <base-logging/Logging.hpp>
-
+#include <fstream>
+#include <traversability_generator3d/SoilNode.hpp>
 #include <yaml-cpp/yaml.h>
 
 #ifdef ENABLE_DEBUG_DRAWINGS
@@ -36,11 +38,13 @@ void TraversabilityGenerator3dGui::setupUI()
     V3DD::CONFIGURE_DEBUG_DRAWINGS_USE_EXISTING_WIDGET(widget);
 #endif
     trav3dViz.setPluginName("TraversabilityMap3d");
+    soil3dViz.setPluginName("SoilMap3d");
     startViz.setPluginName("Start Pose");
 
     widget->setCameraManipulator(vizkit3d::ORBIT_MANIPULATOR);
     widget->addPlugin(&mlsViz);
     widget->addPlugin(&trav3dViz);
+    widget->addPlugin(&soil3dViz);
     widget->addPlugin(&startViz);
     widget->addPlugin(&gridViz);
     
@@ -55,7 +59,45 @@ void TraversabilityGenerator3dGui::setupUI()
     QVBoxLayout* layout = new QVBoxLayout();
     layout->addWidget(widget);
     
-    resetButton = new QPushButton("Reset Traversability Map");
+    QHBoxLayout* soilLayout = new QHBoxLayout();
+
+    soilTypeCombo = new QComboBox();
+    soilTypeCombo->addItem("No selection");   // index 0
+    soilTypeCombo->addItem("CONCRETE");
+    soilTypeCombo->addItem("ROCKS");
+    soilTypeCombo->addItem("SAND");
+    soilTypeCombo->addItem("GRAVEL");
+
+    sigmaXSpin = new QDoubleSpinBox();
+    sigmaXSpin->setRange(0.01, 100.0);
+    sigmaXSpin->setValue(1.0);
+    sigmaXSpin->setSingleStep(0.1);
+    sigmaXSpin->setPrefix("σx=");
+
+    sigmaYSpin = new QDoubleSpinBox();
+    sigmaYSpin->setRange(0.01, 100.0);
+    sigmaYSpin->setValue(1.0);
+    sigmaYSpin->setSingleStep(0.1);
+    sigmaYSpin->setPrefix("σy=");
+
+    uncertaintySpin = new QDoubleSpinBox();
+    uncertaintySpin->setRange(0.0, 1.0);
+    uncertaintySpin->setValue(0.0);
+    uncertaintySpin->setSingleStep(0.05);
+    uncertaintySpin->setPrefix("u=");
+
+    soilHintLabel = new QLabel("Click map to add soil sample");
+
+    soilLayout->addWidget(new QLabel("Soil:"));
+    soilLayout->addWidget(soilTypeCombo);
+    soilLayout->addWidget(sigmaXSpin);
+    soilLayout->addWidget(sigmaYSpin);
+    soilLayout->addWidget(uncertaintySpin);
+
+    layout->addLayout(soilLayout);
+    layout->addWidget(soilHintLabel);
+
+    resetButton = new QPushButton("Reset Traversability and Soil Maps");
     resetButton->setEnabled(false);
     layout->addWidget(resetButton);
     window.setLayout(layout);
[…]
