### 2025-12-30 dfki-ric/traversability_generator3d#13 @haider8645
We added interactive soil sampling to the GUI, allowing users to inject soil semantics directly into the map:

- A soil type dropdown (CONCRETE, ROCKS, SAND, GRAVEL, No selection)
- User-configurable Gaussian spread parameters (σx, σy)
- An uncertainty parameter in [0,1]
- Left-click interaction to place soil samples on the MLS map

Each user-added soil sample is internally represented as a SoilSample and propagated using a Gaussian model over the map. The soil information is then fused into the SoilMap3d and affects the resulting TraversabilityMap3d according to the configured robot capabilities (e.g., whether sand, gravel, etc. are traversable).

How to reproduce

- Run the GUI with one of the provided PLY examples
- Expand the traversability map by clicking on the MLS surface
- Select a soil type and parameters in the GUI
- Click on the map to add soil samples
- Observe how soil semantics influence both the soil map and the traversability map

This workflow is now documented in the README under “Interactive GUI”.
