### 2026-01-02 dfki-ric/traversability_generator3d#13 @haider8645
I would like to follow up specifically on the points related to the example executable, tests, and documentation, as we have made several updates in this regard.

### Walkthrough executable (`travgen3d_walkthrough`)
We clarified in the README that the walkthrough executable is intentionally a **minimal reference example**, not a benchmarking or evaluation tool. Its purpose is to demonstrate the typical API usage of `TraversabilityGenerator3d` on a real point cloud (loading a PLY file, building the MLS map, initializing the generator, and expanding the traversability map).

Importantly, the walkthrough currently **assumes a fixed start position at `(0, 0, -distToGround)`**. If the point cloud origin is not located on a traversable surface (as in your “two cliffs” example), expansion will silently terminate after creating a single node. This behavior is now explicitly documented in the README, together with a note that users should adapt the start position in the code for their own datasets.

### Tests
We have added a dedicated test suite under `test/` and documented it in the README. The unit tests are based on Boost.Test and operate on synthetic MLS maps with controlled geometry. They cover, among others:

- Traversability map generation from MLS input
- Step-height handling and obstacle detection
- Frontier detection and frontier inflation

The tests can be built and run via:
```bash
./build/test/test_TraversabilityGenerator3d
