### 2026-01-03 dfki-ric/traversability_generator3d#13 @haider8645
ROS / ROS 2 and RViz visualization

The ROS 2 visualizations for the traversability maps are currently provided in the [ugv_nav4d_ros2](https://github.com/dfki-ric/ugv_nav4d_ros2) package, rather than directly in traversability_generator3d. This design choice was intentional: the traversability generator itself is a backend library, while the planner (ugv_nav4d) is the component that actively consumes the traversability map and therefore required tight, task-specific RViz visualizations.

Having the visualization as a fully standalone component inside traversability_generator3d would make limited sense without a concrete consumer. Since the planner already uses the traversability maps for motion planning, we developed and validated the RViz visualization there.

Other users have two options:

1. Directly use the planner (ugv_nav4d_ros2) together with its existing RViz visualization for traversability maps, or

1. Use traversability_generator3d as a standalone library and implement custom visualizations tailored to their own ROS / ROS 2 pipeline.

We agree that this relationship was not sufficiently documented. We will clarify this design decision in the documentation and explicitly reference the ROS 2 visualization entry point in ugv_nav4d_ros2. Examples for minimal usage can be found in the documentation of [ugv_nav4d](https://github.com/dfki-ric/ugv_nav4d)
