### 2026-01-19 #9410 @boisgera
A friendly nudge @bstanciulescu, @c-joly 😊

> I have seen several updates by [@haider8645](https://github.com/haider8645) in the project meta-issue opened by [@c-joly](https://github.com/c-joly) ([dfki-ric/traversability_generator3d#13](https://github.com/dfki-ric/traversability_generator3d/issues/13))
>
> [@bstanciulescu](https://github.com/bstanciulescu), [@c-joly](https://github.com/c-joly) could you have a look and possibly update your review checklists accordingly?

Tell me if you need anything to move forward!
