### 2026-01-27 #9410 @haider8645
Hi @boisgera, @c-joly, @bstanciulescu,

thanks again for the follow-up and nudges. Just to confirm on my side: I’ve addressed the points raised so far in the meta-issue (dfki-ric/traversability_generator3d#13), including documentation, installation notes, and examples.

No rush at all. Happy to clarify or adjust anything whenever you have time to resume the review.

Many thanks!
