### 2026-02-10 dfki-ric/traversability_generator3d#13 @c-joly
Hi @haider8645,
Cc: @boisgera, @bstanciulescu

I have checked all your comments and modifications to the code. Everything is fine for me and I will validate the review on JOSS. Just two minor remarks:
   1. I regret a little bit that the function in the GUI to add a soil layer is linked to the function to generate the traversability map. In the current implementation, I can add only one area with a specific soil. I may prefer to have a function to add as many specific soil areas as I want (in that case, each click only generate one extra soil layer and do not start the traversability computation) and an other separate function to launch the traversability computation. If you can change the GUI application to support this functionality, I think it can be a great enhancement.
   2. All my comments are linked to the `paper` branch of the project. After the paper acceptance, do you plan to push the updated codes and doc to the `main` branch?
