### 2026-02-10 #9410 @c-joly
Hi @haider8645
Cc: @boisgera, @bstanciulescu

I have finished my review. In my opinion, the project fulfill all the requirements to be published in JOSS; great work and thank you very much for addressing my (many) bullet points!
