### 2026-02-11 45a1ce8
Minor updates to soil node handling
