### 2026-02-11 #9410 @boisgera
Thanks @c-joly, @bstanciulescu for your reviews! 🙏

Let's start the "post-review" process (step 11 in https://joss.readthedocs.io/en/latest/editing.html).
