### 2026-02-11 #9410 @boisgera
Hi @haider8645,

Could you check if the DOIs suggested by our editorial bot are valid and, if so, add them to you bibliography?

- 10.1109/iros.2006.282632 may be a valid DOI for title: Multi-level surface maps for outdoor terrain mappi...
- 10.1007/s10514-012-9321-0 may be a valid DOI for title: OctoMap: An efficient probabilistic 3D mapping fra...
- 10.1109/lra.2018.2849506 may be a valid DOI for title: Probabilistic terrain mapping for mobile robots wi...
- 10.1126/scirobotics.adj9699 may be a valid DOI for title: Cooperative robotic exploration of a planetary sky...
