+++ paper/paper.bib
@@ -1,5 +1,6 @@
 @inproceedings{triebel2006multilevel,
   title={Multi-level surface maps for outdoor terrain mapping and loop closing},
+  doi={10.1109/iros.2006.282632},
   author={Triebel, Rudolph and Pfaff, Patrick and Burgard, Wolfram},
   booktitle={2006 IEEE/RSJ International Conference on Intelligent Robots and Systems},
   pages={2276--2282},
@@ -22,6 +23,7 @@
 
 @article{hornung2013octomap,
   title={OctoMap: An efficient probabilistic 3D mapping framework based on octrees},
+  doi={10.1007/s10514-012-9321-0},
   author={Hornung, Armin and Wurm, Kai M and Bennewitz, Maren and Stachniss, Cyrill and Burgard, Wolfram},
   journal={Autonomous Robots},
   volume={34},
@@ -33,6 +35,7 @@
 
 @inproceedings{fankhauser2018probabilistic,
   title={Probabilistic terrain mapping for mobile robots with uncertain localization},
+  doi={10.1109/lra.2018.2849506},
   author={Fankhauser, Péter and Bloesch, Michael and Rodriguez, Daniel and Kaestner, Markus and Hutter, Marco and Siegwart, Roland},
   booktitle={2018 IEEE International Conference on Robotics and Automation (ICRA)},
   pages={1875--1882},
@@ -81,6 +84,7 @@
 
 @article{dominguez2025cooperative,
   title={Cooperative robotic exploration of a planetary skylight surface and lava cave},
+  doi={10.1126/scirobotics.adj9699},
   author={Dom{\'\i}nguez, Ra{\'u}l and P{\'e}rez-del-Pulgar, Carlos and Paz-Delgado, Gonzalo J and Polisano, Fabio and Babel, Jonathan and Germa, Thierry and Dragomir, Iulia and Ciarletti, Val{\'e}rie and Berthet, Anne-Claire and Danter, Leon Cedric and others},
   journal={Science Robotics},
   volume={10},
