### 2026-02-11 #9410 @haider8645
@boisgera Thanks, I have added the missing DOIs.

Also created an archive on Zenodo here: https://doi.org/10.5281/zenodo.18611473
