### 2026-02-12 dfki-ric/traversability_generator3d#13 @haider8645
1. I checked the code and you are spot about the traversability computation on every newly added soil sample. I do have a solution in mind, as you suggested, to collect soil samples without any traversability computation. Can we move the point 1 to a new issue so that this issue can be closed?
2. Yes, I will merge everything from paper to main.
