### 2026-02-12 dfki-ric/traversability_generator3d#13 @c-joly
Hello @haider8645
You can close this issue now. I added https://github.com/dfki-ric/traversability_generator3d/issues/14 in which a copied-pasted the content of point 1.
Best,
