### 2026-02-24 #9410 @crvernon
@openjournals/dev I don't see a specific reason of why acceptance is failing.  Could you provide insight here?  Thanks.
