### 2026-02-24 #9410 @haider8645
@crvernon Thanks!

The DOI page has an error. I assume this will go away later?

```
404
File not found

The site configured at this address does not contain the requested file.

If this is your site, make sure that the filename case matches the URL as well as any file permissions.
For root URLs (like http://example.com/) you must provide an index.html file.

[Read the full documentation](https://help.github.com/pages/) for more information about using GitHub Pages.

[GitHub Status](https://githubstatus.com/) — [@githubstatus](https://twitter.com/githubstatus)
```
