### 2026-02-24 416a242
Create CITATION.cff
