+++ README.md
@@ -2,6 +2,9 @@
 
 CI build (paper): ![paper](https://github.com/dfki-ric/traversability_generator3d/actions/workflows/c-cpp.yml/badge.svg)
 
+JOSS Paper: 
+[![DOI](https://joss.theoj.org/papers/10.21105/joss.09410/status.svg)](https://doi.org/10.21105/joss.09410)
+
 [![License: BSD-3-Clause](https://img.shields.io/badge/License-BSD%203--Clause-blue.svg)](LICENSE)
 
 A C++ library for **3D traversability estimation from Multi-Level Surface (MLS) maps**, developed at the [DFKI Robotics Innovation Center](https://robotik.dfki-bremen.de/en/startpage.html).
