### 2026-02-24 #9410 @chhtz
> The DOI page has an error. I assume this will go away later?

Page works for me (maybe it needed 5 more minutes ^^)
